refactor(shared,preload): domain-split the ipc contract + one preload subscribe helper
quality-review slice 5 (5c, 5d).

5c: split the 1026-line src/shared/ipc.ts into src/shared/ipc/ by domain
(core / auth / thread / git / gh / files), each module owning BOTH its channel
entries and its Args/Result types so the wire name and its payload docs live
together. A barrel (src/shared/ipc/index.ts) assembles the single `IPC` const from
the per-domain `as const` channel fragments (keys + string values byte-identical to
before — verified against HEAD) and re-exports every type, so every existing
`import … from '../shared/ipc'` resolves unchanged via the directory index. Killed the
double documentation: each channel's deep prose now lives once (on its type), with a
one-line pointer on the channel entry; unique content merged, none lost. The directory
stays free of Node/DOM imports.

5d: collapse the six byte-identical `on*` bridges in preload to a local
`subscribe<T>(channel)` helper; each `on*` is now `subscribe<Payload>(IPC.x)`.
